% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddRowMeansToFeature.R
\name{AddRowMeansToFeature}
\alias{AddRowMeansToFeature}
\title{Add Row Mean from Data Table to Feature Table}
\usage{
AddRowMeansToFeature(data_table, features)
}
\arguments{
\item{data_table}{A `data.table` with a `row.names` column and numeric columns for which row mean should be calculated.}

\item{features}{A `data.table` with a `V1` column corresponding to feature IDs (matching `row.names` in `data_table`).}
}
\value{
A `data.table` containing all columns from `features` and a new `sum` column with the row means from `data_table`.
}
\description{
Calculates the row Mean of numeric columns in a data table and merges these mean into the features table.
}
