% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddRowSumToFeature.R
\name{AddRowSumToFeature}
\alias{AddRowSumToFeature}
\title{Add Row Sums from Data Table to Feature Table}
\usage{
AddRowSumToFeature(data_table, features)
}
\arguments{
\item{data_table}{A `data.table` with a `row.names` column and numeric columns for which row sums should be calculated.}

\item{features}{A `data.table` with a `V1` column corresponding to feature IDs (matching `row.names` in `data_table`).}
}
\value{
A `data.table` containing all columns from `features` and a new `sum` column with the row sums from `data_table`.
}
\description{
Calculates the row sums of numeric columns in a data table and merges these sums into the features table.
}
