% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneStats.R
\name{GeneStats}
\alias{GeneStats}
\title{Calculate Gene Annotation Statistics}
\usage{
GeneStats(expression, atac)
}
\arguments{
\item{expression}{A  GRanges object with a `Gene_Expression_id` column for gene expression features.}

\item{atac}{A GRanges object with a `gene_id` column for ATAC-seq features.}
}
\value{
A data.frame with one row and the following columns:
  - `tot_genes`: Total number of genes in the expression data
  - `merged_num`: Number of genes with at least one overlapping ATAC peak
  - `non_merged`: Number of genes without any overlapping ATAC peak
  - `prc_merged`: Percentage of merged genes out of total
  - `prc_unmerged`: Percentage of non-merged genes out of total
}
\description{
Computes statistics on the overlap between gene expression features and ATAC-seq peaks.
}
