% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogCPMCombinedPlot.R
\name{LogCPMCombinedPlot}
\alias{LogCPMCombinedPlot}
\title{Plot Mean ATAC vs. RNA Values by Chromosome}
\usage{
LogCPMCombinedPlot(
  atac_data,
  expression_data,
  atac_features,
  expression_features,
  gr_atac
)
}
\arguments{
\item{atac_data}{A data frame or data.table of normalized ATAC values, with
a `row.names` column containing peak IDs.}

\item{expression_data}{A data frame or data.table of normalized expression values,
with a `row.names` column containing gene IDs.}

\item{atac_features}{A data frame or data.table of ATAC feature metadata. Must
have a `V1` column containing peak IDs.}

\item{expression_features}{A data frame or data.table of expression feature metadata.
Must have a `V1` column containing gene IDs.}

\item{gr_atac}{A `GRanges` object of annotated ATAC-seq peaks. Must have
metadata columns `gene_id` and `Peaks_id`.}
}
\value{
A named list of `ggplot` objects, where each element corresponds to a
  chromosome and contains the scatter plot.
}
\description{
For each chromosome, this function generates a scatter plot comparing the mean
normalized ATAC-seq values against the mean normalized RNA-seq values for
features linked by a common gene.
}
