% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotATACDistribution.R
\name{PlotATACDistribution}
\alias{PlotATACDistribution}
\title{Plot ATAC-seq Peak Intensity Distribution for Top Chromosomes}
\usage{
PlotATACDistribution(gr_atac)
}
\arguments{
\item{gr_atac}{A `GRanges` object for ATAC-seq peaks, with metadata column `Peaks_sum` (peak intensity).}
}
\value{
A named list of `ggplot2` objects, one per chromosome, visualizing peak intensity by position and size.
}
\description{
Generates scatter plots of ATAC-seq peak intensity distribution across genomic positions (chromosomes in a GRanges object.)
}
